refactor: command structure
