+++ commands/add/index.js
@@ -11,5 +11,19 @@ module.exports = (program) => {
         .help('A simple alias to `yarn add` command.')
         .option('<module1> <module2> <module3>', 'The modules to add')
         .option('[--dev]', 'Add to dev dependencies.')
-        .action(`${__dirname}/action.js`);
+        .action(async (app, options) => {
+            const colors = require('colors/safe');
+            const Project = require('../../lib/Project');
+
+            const cwd = process.cwd();
+            const project = new Project(cwd);
+
+            // Add requested packages.
+            if (options.dev) {
+                return await project.packageManager.dev(...options.arguments);
+            }
+
+            await project.packageManager.add(...options.arguments);
+            app.log(colors.green('packages successfully added.'));
+        });
 };
