+++ commands/publish/index.js
@@ -13,5 +13,29 @@ module.exports = (program) => {
         .option('[--no-git]', 'Do not commit version changes to Git.')
         .option('[--no-npm]', 'Do not commit version changes to NPM.')
         .help('Use `lerna` to handle to publish monorepos.\n')
-        .action(`${__dirname}/action.js`);
+        .action(async (app, options) => {
+            const exec = require('../../lib/exec.js');
+            const BIN = require.resolve('lerna/cli.js');
+
+            let args = ['publish', '--use-workspaces'];
+            if (options.beta) {
+                args.push('--canary');
+                if (!options.hasOwnProperty('exact')) {
+                    options.exact = true;
+                }
+            }
+            if (options.exact) {
+                args.push('--exact');
+            }
+            if (options.git === false) {
+                args.push('--skip-git');
+            }
+            if (options.npm === false) {
+                args.push('--skip-npm');
+            }
+            if (process.env.CI) {
+                args.push('--yes');
+            }
+            return await exec(BIN, args);
+        });
 };
