+++ commands/start/index.js
@@ -9,5 +9,12 @@ module.exports = (program) => {
         .command('start')
         .description('Trigger project `start` script.')
         .help('A simple alias to `yarn start` command.')
-        .action(`${__dirname}/action.js`);
+        .action(async () => {
+            const Project = require('../../lib/Project');
+
+            const cwd = process.cwd();
+            const project = new Project(cwd);
+
+            return await project.packageManager.start();
+        });
 };
