+++ commands/test/index.js
@@ -9,5 +9,12 @@ module.exports = (program) => {
         .command('test')
         .description('Trigger project `test` script.')
         .help('A simple alias to `yarn test` command.')
-        .action(`${__dirname}/action.js`);
+        .action(async () => {
+            const Project = require('../../lib/Project');
+
+            const cwd = process.cwd();
+            const project = new Project(cwd);
+
+            return await project.packageManager.test();
+        });
 };
