--- configs/browserslist.json
@@ -1,12 +1,0 @@
-[
-    "ie >= 11",
-    "last 3 iOS major versions",
-    "Android >= 4.4",
-    "last 3 Safari major versions",
-    "last 3 Firefox major versions",
-    "unreleased Firefox versions",
-    "Chrome 45",
-    "last 3 Chrome major versions",
-    "unreleased Chrome versions",
-    "last 3 Edge major versions"
-]
