+++ lib/Bundlers/plugins/sass-resolver/sass-resolver.js
@@ -1,6 +1,6 @@
 const fs = require('fs-extra');
 const path = require('path');
-const ext = require('../../../../lib/extensions.js');
+const { STYLE_EXTENSIONS, isStyleFile } = require('../../../../lib/extensions.js');
 
 /**
  * Generate a list of file paths with all style extensions.
@@ -12,7 +12,7 @@ function alternatives(url) {
         // url already has an extension.
         [url] :
         // remap the path with all style extensions.
-        ext.STYLE_EXTENSIONS.map((ext) => `${url}${ext}`);
+        STYLE_EXTENSIONS.map((ext) => `${url}${ext}`);
     // look for sass partials too.
     if (path.basename(url)[0] !== '_') {
         for (let i = 0, len = res.length; i < len; i++) {
@@ -82,7 +82,7 @@ module.exports = function() {
                         if (json.style) {
                             // style field found.
                             url = path.join(path.dirname(checked), json.style);
-                        } else if (json.main && ext.isStyleFile(json.main)) {
+                        } else if (json.main && isStyleFile(json.main)) {
                             // try to use the main field if it is a css file.
                             url = path.join(path.dirname(checked), json.main);
                         }
