+++ lib/Navigator.js
@@ -1,6 +1,7 @@
 const fs = require('fs-extra');
 const path = require('path');
 const glob = require('glob');
+const gzipSize = require('gzip-size');
 
 class NavigatorEntry {
     constructor(file, cwd) {
@@ -16,11 +17,26 @@ class NavigatorEntry {
         return path.extname(this.path);
     }
 
+    get dirname() {
+        return path.dirname(this.path);
+    }
+
+    get size() {
+        return {
+            size: fs.statSync(this.path).size,
+            zipped: gzipSize.sync(this.path),
+        };
+    }
+
     get localPath() {
         if (this.cwd) {
             return path.relative(this.cwd, this.path);
         }
         return this.path;
+    }
+
+    relative(file) {
+        return path.relative(this.path, file);
     }
 
     isFile() {
@@ -43,6 +59,10 @@ class NavigatorEntry {
 }
 
 class NavigatorFile extends NavigatorEntry {
+    get mapFile() {
+        return new NavigatorFile(`${this.path}.map`, this.cwd || this.path);
+    }
+
     read() {
         try {
             return fs.readFileSync(this.path, 'utf8');
@@ -70,8 +90,12 @@ class NavigatorFile extends NavigatorEntry {
 }
 
 class NavigatorDirectory extends NavigatorEntry {
-    ensureExists() {
+    ensure() {
         fs.ensureDirSync(this.path);
+    }
+
+    empty() {
+        fs.emptyDirSync(this.path);
     }
 
     resolve(patterns) {
@@ -111,6 +135,37 @@ class NavigatorDirectory extends NavigatorEntry {
 
     directory(directory) {
         return new NavigatorDirectory(path.resolve(this.path, directory), this.cwd || this.path);
+    }
+
+    children() {
+        if (!this.exists()) {
+            return null;
+        }
+
+        let children = fs.readdirSync(this.path);
+        return children.map((file) => {
+            let entry = this.entry(file);
+            if (entry.isDirectory()) {
+                return this.directory(file);
+            }
+            return this.file(file);
+        });
+    }
+
+    files() {
+        if (!this.exists()) {
+            return null;
+        }
+
+        return this.children().filter((entry) => entry.isFile());
+    }
+
+    directories() {
+        if (!this.exists()) {
+            return null;
+        }
+
+        return this.children().filter((entry) => entry.isDirectory());
     }
 }
