+++ lib/Project.js
@@ -1,11 +1,26 @@
 const path = require('path');
 const { NavigatorDirectory } = require('./Navigator');
+const PackageManager = require('./PackageManager');
 const { keypath } = require('@chialab/proteins');
+
+const DEFAULTS = [
+    'ie >= 11',
+    'last 3 iOS major versions',
+    'Android >= 4.4',
+    'last 3 Safari major versions',
+    'last 3 Firefox major versions',
+    'unreleased Firefox versions',
+    'Chrome 45',
+    'last 3 Chrome major versions',
+    'unreleased Chrome versions',
+    'last 3 Edge major versions',
+];
 
 class Project extends NavigatorDirectory {
     constructor(directory) {
         super(directory);
         this.packageJson = this.file('package.json');
+        this.packageManager = new PackageManager(directory);
         if (this.packageJson.exists()) {
             this.load();
         } else {
@@ -27,7 +42,7 @@ class Project extends NavigatorDirectory {
         return !this.packageJson.exists();
     }
 
-    get hasWorkspace() {
+    get parent() {
         let paths = this.path.split(path.sep).slice(0, -1);
         while (paths.length) {
             paths.pop();
@@ -41,7 +56,7 @@ class Project extends NavigatorDirectory {
             }
             return workspaces.find((ws) => ws.path === this.path);
         }
-        return false;
+        return null;
     }
 
     get directories() {
@@ -65,6 +80,23 @@ class Project extends NavigatorDirectory {
         return directories
             .filter((entry) => entry instanceof NavigatorDirectory)
             .map((entry) => new Project(entry.path));
+    }
+
+    get browserslist() {
+        if (this.file('browserslist.json').exists()) {
+            return this.file('browserslist.json').readJson();
+        }
+
+        if (this.get('browserslist')) {
+            return this.get('browserslist');
+        }
+
+        let parent = this.parent;
+        if (parent) {
+            return parent.browserslist;
+        }
+
+        return DEFAULTS;
     }
 
     save() {
