+++ lib/cli.js
@@ -229,12 +229,7 @@ class CLI {
         }
 
         try {
-            let fn = command.callback;
-            if (typeof fn !== 'function') {
-                fn = require(command.callback);
-            }
-
-            let res = await fn(this, options);
+            const res = await command.callback(this, options);
             profile.end();
             return res;
         } catch (err) {
