+++ lib/utils.js
@@ -1,6 +1,4 @@
-const fs = require('fs-extra');
 const path = require('path');
-const gzipSize = require('gzip-size');
 
 /**
  * @typedef {Object} FileSize
@@ -69,27 +67,6 @@ const UTILS = {
      */
     relativeToCwd(file) {
         return path.relative(process.cwd(), file);
-    },
-
-    /**
-     * Get the file size (original and gzipped).
-     *
-     * @param {string} file The file to analyze.
-     * @return {FileSize|Array<FileSize>}
-     */
-    size(file) {
-        let stats = fs.statSync(file);
-        if (stats.isDirectory()) {
-            return fs.readdirSync(file)
-                .map((child) => path.join(file, child))
-                .filter((child) => fs.statSync(child).isFile())
-                .map((child) => UTILS.size(child));
-        }
-        return {
-            file,
-            size: fs.statSync(file).size,
-            zipped: gzipSize.fileSync(file),
-        };
     },
 
     /**
